refactor: 컨트롤러에 웹 요청/응 책임만 가지도록 리팩토링
